Refactor patient components: move to admin directory and implement new styles
