+++ client/src/components/admin/Patient/PatientProfile.css
@@ -1,0 +1,26 @@
+/* PatientProfile.css - scoped to patient profile component */
+.patient-profile-container {
+  background: #fff;
+  padding: 30px 25px;
+  border-radius: 12px;
+  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
+  max-width: 600px;
+  margin: 50px auto;
+  font-family: 'Arial', sans-serif;
+  color: #333;
+}
+
+.patient-profile-container h2 {
+  text-align: center;
+  margin-bottom: 25px;
+  color: #007bff;
+}
+
+.patient-profile-container p {
+  font-size: 16px;
+  margin-bottom: 15px;
+}
+
+.patient-profile-container b {
+  color: #555;
+}
