+++ client/src/components/admin/Patient/PatientProfile.jsx
@@ -1,5 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
+import './PatientProfile.css';
+
 
 function PatientProfile() {
   const [user, setUser] = useState(null);
@@ -17,13 +19,14 @@ function PatientProfile() {
   if (!user) return <p>Loading profile...</p>;
 
   return (
-    <div style={{ maxWidth: "600px", margin: "auto", padding: "20px" }}>
-      <h2>Welcome, {user.name}</h2>
-      <p><b>Email:</b> {user.email}</p>
-      <p><b>Role:</b> {user.role}</p>
-      {/* Add more patient fields here if needed */}
-    </div>
-  );
+  <div className="patient-profile-container">
+    <h2>Welcome, {user.name}</h2>
+    <p><b>Email:</b> {user.email}</p>
+    <p><b>Role:</b> {user.role}</p>
+    {/* Add more patient fields here if needed */}
+  </div>
+);
+
 }
 
 export default PatientProfile;
