+++ client/src/components/admin/Patient/Register.jsx
@@ -1,0 +1,64 @@
+import React, { useState } from "react";
+import axios from "axios";
+import { useNavigate } from "react-router-dom";
+import './Register.css';
+
+
+
+function Register() {
+  const [name, setName] = useState("");
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [msg, setMsg] = useState("");
+
+    const navigate = useNavigate();
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    try {
+      const res = await axios.post("http://localhost:7000/api/auth/register", {
+        name,
+        email,
+        password,
+      });
+      setMsg(res.data.message);
+      navigate("/");
+    } catch (err) {
+      setMsg(err.response?.data?.message || "Error registering user");
+    }
+  };
+
+  return (
+  <div className="register-container">
+    <h2>Sign Up</h2>
+    <form onSubmit={handleSubmit}>
+      <input
+        type="text"
+        placeholder="Enter name"
+        value={name}
+        onChange={(e) => setName(e.target.value)}
+        required
+      /><br /><br />
+      <input
+        type="email"
+        placeholder="Enter email"
+        value={email}
+        onChange={(e) => setEmail(e.target.value)}
+        required
+      /><br /><br />
+      <input
+        type="password"
+        placeholder="Enter password"
+        value={password}
+        onChange={(e) => setPassword(e.target.value)}
+        required
+      /><br /><br />
+      <button type="submit">Sign Up</button>
+    </form>
+    <p>{msg}</p>
+  </div>
+);
+
+}
+
+export default Register;
